feat: updated profile form
